requirements.txt
# Pin versions of following dependencies:
setuptools>=38.3.0
cython>=0.29
numpy>=1.16
scipy==1.2.1 # Statsmodels 0.9/Scipy 1.3.0 incompatibility
dask[complete]==1.2.2 # for future development
pandas==0.24.0
pystan==2.19.0.0
joblib==0.13.2
statsmodels==0.9.0
biopython==1.73
pysam==0.15.2 # with Samtools-1.7
pyyaml==5.1
